+++ include/ucoind.h
@@ -237,11 +237,18 @@ typedef struct {
  * prototypes
  ********************************************************************/
 
+const uint8_t *ucoind_nodeid(void);
+
 bool forward_payment(fwd_proc_add_t *p_add);
 bool backward_fulfill(const ln_cb_fulfill_htlc_recv_t *p_fulfill);
 bool backward_fail(const ln_cb_fail_htlc_recv_t *pFail);
+
 void preimage_lock(void);
 void preimage_unlock(void);
+
+bool monitering_get(void);
+void monitering_stop(void);
+
 bool close_unilateral_local(ln_self_t *self, void *pDbParam);
 
 #endif /* UCOIND_H__ */
