+++ ucoind/inc/cmd_json.h
@@ -1,0 +1,8 @@
+﻿#ifndef CMD_JSON_H__
+#define CMD_JSON_H__
+
+#include <inttypes.h>
+
+void cmd_json_start(uint16_t Port);
+
+#endif  //CMD_JSON_H__
